+++ src/settings.ts
@@ -1,0 +1,24 @@
+import { SlideImprovementsSettings } from "./types";
+
+export const DEFAULT_SETTINGS: SlideImprovementsSettings = {
+  enableStyling: true,
+  outputFolder: "",
+  baseFontSize: "1.6em",
+  h1FontSize: "2em",
+  h2FontSize: "1.4em",
+  slidePadding: "3vw",
+  headingMarginTop: "2.5em",
+  accentColor: "#A2CF80",
+  scrollableSlides: true,
+  h1Color: "#A2CF80",
+  h2Color: "#FFD700",
+  h3Color: "#FF8C00",
+  h4Color: "#1E90FF",
+  h5Color: "#BA55D3",
+  h6Color: "#FF69B4",
+  enableMobileStyling: true,
+  mobileFontSizeVertical: "4vw",
+  mobileFontSizeHorizontal: "3vw",
+  mobileScrollableSlides: true,
+  centerMobileVertically: false
+};
