+++ src/types.ts
@@ -1,0 +1,22 @@
+export interface SlideImprovementsSettings {
+  enableStyling: boolean;
+  outputFolder: string;
+  baseFontSize: string;
+  h1FontSize: string;
+  h2FontSize: string;
+  slidePadding: string;
+  headingMarginTop: string;
+  accentColor: string;
+  scrollableSlides: boolean;
+  h1Color: string;
+  h2Color: string;
+  h3Color: string;
+  h4Color: string;
+  h5Color: string;
+  h6Color: string;
+  enableMobileStyling: boolean;
+  mobileFontSizeVertical: string;
+  mobileFontSizeHorizontal: string;
+  mobileScrollableSlides: boolean;
+  centerMobileVertically: boolean;
+}
