+++ src/ui/SlideImprovementsSettingTab.ts
@@ -1,0 +1,7 @@
+import { PluginSettingTab, App, Setting } from "obsidian";
+import { SlideImprovementsSettings } from "../types";
+import { ObsidianSlideImprovementsPlugin } from "../main"; // adjust import as needed
+
+export class SlideImprovementsSettingTab extends PluginSettingTab {
+  // ...full class code...
+}
