+++ src/utils/injectCss.ts
@@ -1,0 +1,11 @@
+import { SlideImprovementsSettings } from "../types";
+
+/** Styling CSS */
+export function injectSlideCss(settings: SlideImprovementsSettings) {
+  // ...full code from before, but only for style CSS...
+}
+
+/** Scroll CSS */
+export function injectScrollCss(settings: SlideImprovementsSettings) {
+  // ...full code from before, but only for scroll CSS...
+}
