+++ tsconfig.json
@@ -6,11 +6,12 @@
     "lib": ["ES2020", "DOM"],
     "declaration": false,
     "outDir": "./",
-    "rootDir": "./",
+    "rootDir": "./src",
     "strict": true,
     "esModuleInterop": true,
-    "skipLibCheck": true
+    "skipLibCheck": true,
+    "noEmit": true
   },
-  "include": ["main.ts"],
+  "include": ["src"],
   "exclude": ["node_modules"]
 }
